refactor(build): extract common function to save token logo (#6031)
# Motivation

To avoid repetition of code, we extract a common function to save the
logo of ICRC and ckERC20 tokens in the build scripts.
