+++ scripts/build.tokens.utils.ts
@@ -5,6 +5,7 @@ import { IcrcLedgerCanister, mapTokenMetadata } from '@dfinity/ledger-icrc';
 import type { IcrcTokenMetadataResponse } from '@dfinity/ledger-icrc/dist/types/types/ledger.responses';
 import { Principal } from '@dfinity/principal';
 import { createAgent } from '@dfinity/utils';
+import { closeSync, openSync, writeSync } from 'node:fs';
 
 export const agent: HttpAgent = await createAgent({
 	identity: new AnonymousIdentity(),
@@ -26,4 +27,24 @@ export const loadMetadata = async (
 	const metadata = await getMetadata(Principal.from(ledgerCanisterId));
 
 	return mapTokenMetadata(metadata);
+};
+
+export const saveLogo = ({ logoData, file }: { logoData: string; file: string }) => {
+	const [encoding, encodedStr] = logoData.split(';')[1].split(',');
+
+	const svgContent = Buffer.from(encodedStr, encoding as BufferEncoding).toString('utf-8');
+
+	try {
+		const fd = openSync(file, 'wx');
+
+		writeSync(fd, svgContent, 0, 'utf-8');
+		closeSync(fd);
+	} catch (err: unknown) {
+		if (typeof err === 'object' && err !== null && 'code' in err && err.code === 'EEXIST') {
+			// File already exists, do nothing
+			return;
+		}
+
+		throw err;
+	}
 };
